Add utils::CopyToClone and use it in providers::core::Provider
Rust currently doesn't allow to clone function pointers with references
as arguments, which causes problems with the Provider struct. In order
to avoid that, all the arguments of the providers are currently owned,
which is fine but wastes a lot of memory.

In order to fix this in the future, this commit introduces
utils::CopyToClone, a wrapper struct which allows to Clone anything
which implements Copy.