Add support for working behind a proxy
Before this commit, there was no way to use Fisher behind a proxy and
rely on the source IP address, because all the IP addresses were the
proxy's one.

In order to work around this, proxies add an X-Forwarded-For HTTP header
to all the request, so the applications can parse it and derive the
correct IP address.

This poses a security problem though. If someone adds the header on his
client, he's able to spoof his address, breaking the security checks. In
order to fix this, proxy support is only available when you provide the
--behind-proxies flag with the number of proxies you're using:

$ fisher hooks-dir/ --behind-proxies 1

Fixes: GH-3