Fix intermittent failure in the processor's shutdown
There was a bug in the threads shutdown code which caused Fisher to
try stopping a non-existing thread, causing a crash.

This was possible because the shutdown code is split in two phases: the
first one checks which threads can be stopped, and then the other one
removes all the threads at the indexes provided by the first step.

But once a thread was stopped - and removed from the threads list - the
remaining indexes to remove weren't decremented by one, and the code
crashed.

This commit fixes the issue.