Refactor status hooks to use their own request kind
Before this commit, status hooks (ab)used the Request::Web kind,
serializing and deserializing the state with JSON. This causes slow
performance due to the serialization/deserialization, but also is more
prone to bugs since everything is controlled by strings, so the compiler
can't check it.

This commit introduces a new request kind, Request::Status, which
carries all the information needed for the status hooks, without any
serialization. Also, the events are represented by an enum, so
everything can be checked at compile time.

Due to this change, $FISHER_REQUEST_BODY is not available anymore on
status hooks. This is a breaking change.