Keep the web server alive while shutting down
Before this commit, the web server was stopped before the processor.
This means if there was a backlog of jobs to be processed, the web
server wasn't available to reply to requests like /health.

This was done because this way no more jobs were added to the queue.

This commit fixes this behavior by stopping the web server after the
processor. This also prevents new jobs to be queued after the shutdown
signal was received, while keeping the web server alive.