Add support for prioritizing jobs in the processor
This commit *doesn't* expose prioritized jobs to the users, but adds
support for them in the processor. This will then be used by status
hooks, and possibly exposed to the users later.

This commit also adds a Serial type, which is used by the processor to
keep the ordering of the queued jobs. A Serial can't overflow, but it
might misbehave if there are more than 2^32 queued jobs.